requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
loguru==0.6.0
python-telegram-bot-calendar==1.0.5
peewee==3.14.10